Add files via upload
Make a Registration  Functionality Using Prompt.
Rules :
Take information from the user (User name, phone Number email, password and confirm password)
The prompt should not be empty. if the user does not enter any value from the prompt and shows an Error please fill in the input
if the user fills in the name phone number email then check the password and confirm the password should be the same
if same shows alert you are successfully registered else password should be same.